requirements.txt
# Python+C++混合架构依赖包
# 竹子识别系统Python组件依赖

# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
pyyaml>=6.0
dataclasses-json>=0.5.7

# 网络通信
pymodbus>=3.0.0
websockets>=10.0

# GUI库(可选)
tkinter-tooltip>=2.0.0
pillow>=8.0.0

# 日志和配置
loguru>=0.6.0
pydantic>=1.8.0

# 系统监控
psutil>=5.8.0
GPUtil>=1.4.0

# 开发工具
pytest>=6.0.0
pytest-asyncio>=0.18.0
black>=21.0.0
flake8>=4.0.0

# C++绑定
pybind11>=2.8.0

# LVGL Python绑定
lvgl>=8.3.0

# 数据处理
pandas>=1.3.0
matplotlib>=3.4.0

# 异步支持
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0

# 性能分析
memory-profiler>=0.60.0
line-profiler>=3.3.0

# 配置文件处理
toml>=0.10.2
configparser>=5.2.0

# 类型检查
mypy>=0.910
types-PyYAML>=6.0.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0